delete api done

+++ src/main/java/edu/icet/employeecrm/repository/EmployeeRepository.java
@@ -2,5 +2,5 @@ package edu.icet.employeecrm.repository;
 import edu.icet.employeecrm.entity.EmployeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<EmployeeEntity,Integer>{
+public interface EmployeeRepository extends JpaRepository<EmployeeEntity,Long>{
 }
